+++ mayan/apps/cabinets/views.py
@@ -6,7 +6,6 @@ from django.urls import reverse_lazy
 from django.utils.translation import ugettext_lazy as _, ungettext
 
 from mayan.apps.acls.models import AccessControlList
-from mayan.apps.documents.permissions import permission_document_view
 from mayan.apps.documents.models import Document
 from mayan.apps.documents.views.document_views import DocumentListView
 from mayan.apps.views.generics import (
@@ -272,7 +271,7 @@ class DocumentCabinetAddView(MultipleObjectFormActionView):
 
 
 class DocumentCabinetListView(ExternalObjectViewMixin, CabinetListView):
-    external_object_permission = permission_document_view
+    external_object_permission = permission_cabinet_view
     external_object_pk_url_kwarg = 'document_id'
     external_object_queryset = Document.valid